The uart_getline function is embedded in an atomic block, preventing another interrupt to abort, that could cause a sequence of calls to this function.
due to problems with stability and reading chars when the register
has not yet parsed the next character in queue, a delay of 0.5 milli-
seconds is added in the polling iteration as to prevent unwanted null
or other garbage data to interfere with the string construction.

Improved parsing algorithm with more readable code, as well as memory
overflow protection.