Assignment 1.4.2:
In the idle_state which is being returned to constantly by the
design of StateMachine.h, the 'staticState' property of the state
machine is updated if the key is pressed. The button is
debounced by it's own defined timeframe and the staticState is updated
if the debouncing meets the requirements of continuity.