The statemachine instance now has a state chain with the following diagram:
PULSE -> POTENTIOMETER_LED -> FLASH_LED -> OFF -> PULSE -> (...)

This allows the idle (main state) state a.k.a. idle_state() to
set the static state to be the next chained state for the given
current state.

Since idle_state() is being returned to automatically by the
statemachine every time, this need to store the next state was
enabled by setting the StaticState property in the statemachine
to the current one.

Upon a button push-and-release event, this static state is mutated
to the next state in line - by calling stateMachine.getChainedStateForState()
which returns the next state in the sequenced declared above.

This will make the statemachine return the function pointer to that
state when the main loop calls next() on the statemachine instance,
thus getting a button-controlled stateful, nonblocking firmware
that can continuously check buttons and other entities in the
idle_state.